move new tests to renamed package, fix epl tck bundle generation
