+++ jaxrs-tck-docs/JAXRSTCK3.1-ReleaseNotes.html
@@ -74,7 +74,7 @@ td {  padding-left: 3pt;}
     <h2>Jakarta EE RESTful Web Services 3.1 TCK Facts</h2>
     <p>The test suite bundle contains the following:</p>
     <ul>
-      <li>2792 total tests</li>
+      <li>2795 total tests</li>
     </ul>
     <hr>
     <h2>Platform Notes</h2>
