+++ jaxrs-tck/src/main/java/ee/jakarta/tck/ws/rs/jaxrs31/spec/extensions/TSAppConfig.java
@@ -14,18 +14,19 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package jakarta.ws.rs.tck.jaxrs31.spec.extensions;
+package ee.jakarta.tck.ws.rs.jaxrs31.spec.extensions;
 
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ApplicationPath;
+import jakarta.ws.rs.core.Application;
 
-@Path("/")
-public class DynamicFeatureResource {
+import java.util.HashSet;
+import java.util.Set;
 
-    @GET
-    @Path("dynamicFeature")
-    public Response getDynamicFeature() {
-        return Response.ok().build();
-    }
+@ApplicationPath("/")
+public class TSAppConfig extends Application {
+  public Set<Class<?>> getClasses() {
+    Set<Class<?>> resources = new HashSet<>();
+    resources.add(DynamicFeatureResource.class);
+    return resources;
+  }
 }
